WEB-3742: moved sessions service to root
