+++ src/authentication/auth-strategies/Local.strategy.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@nestjs/common'
 import { PassportStrategy } from '@nestjs/passport'
 import { Strategy } from 'passport-local'
-import { AuthenticationService } from '../services/authentication.service'
+import { AuthenticationService } from '../authentication.service'
 import {
   LoginPayload,
   LoginPayloadSchema,
