+++ src/authentication/authentication.service.ts
@@ -7,9 +7,9 @@ import {
   UnauthorizedException,
 } from '@nestjs/common'
 import { JsonWebTokenError, JwtService, TokenExpiredError } from '@nestjs/jwt'
-import { UsersService } from '../../users/services/users.service'
-import { CreateUserInputDto } from '../../users/schemas/CreateUserInput.schema'
-import { LoginPayload } from '../schemas/LoginPayload.schema'
+import { UsersService } from '../users/services/users.service'
+import { CreateUserInputDto } from '../users/schemas/CreateUserInput.schema'
+import { LoginPayload } from './schemas/LoginPayload.schema'
 import { compare } from 'bcrypt'
 import { PrismaClientKnownRequestError } from '@prisma/client/runtime/library'
 import { nanoid } from 'nanoid'
@@ -17,7 +17,7 @@ import {
   SocialAuthPayload,
   SocialProvider,
   SocialTokenValidator,
-} from '../authentication.types'
+} from './authentication.types'
 import { OAuth2Client, TokenInfo } from 'google-auth-library'
 
 const googleOAuthClient = new OAuth2Client(process.env.GOOGLE_CLIENT_ID)
